refactor: NotificationService
