+++ packages/shared/src/services/index.ts
@@ -1,2 +1,3 @@
+export * from './NotificationService.js';
 export * from './TranscriptionService.js';
 export * from './TranscriptionServiceWhisperingLive.js';
